feat: colors from types
