+++ src/app/ui/index.css
@@ -9,3 +9,12 @@
 .fc-theme-standard th {
   border: 2px solid #fff;
 }
+
+::-webkit-scrollbar {
+    width: 0;  /* Remove scrollbar space */
+    background: transparent;  /* Optional: just make scrollbar invisible */
+}
+/* Optional: show position indicator in red */
+::-webkit-scrollbar-thumb {
+    background: #FF0000;
+}
